Renames old projects components
